quilkin_proto::generated::quilkin::relay::v1alpha1::aggregated_control_plane_discovery_service_server
Trait AggregatedControlPlaneDiscoveryService

pub trait AggregatedControlPlaneDiscoveryService:
    Send
    + Sync
    + 'static {
    type StreamAggregatedResourcesStream: Stream<Item = Result<DiscoveryRequest, Status>> + Send + 'static;
    type DeltaAggregatedResourcesStream: Stream<Item = Result<DeltaDiscoveryRequest, Status>> + Send + 'static;

    // Required methods
    fn stream_aggregated_resources<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<DiscoveryResponse>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamAggregatedResourcesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delta_aggregated_resources<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<DeltaDiscoveryResponse>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DeltaAggregatedResourcesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with AggregatedControlPlaneDiscoveryServiceServer.

Required Associated Types

type StreamAggregatedResourcesStream: Stream<Item = Result<DiscoveryRequest, Status>> + Send + 'static

Server streaming response type for the StreamAggregatedResources method.

type DeltaAggregatedResourcesStream: Stream<Item = Result<DeltaDiscoveryRequest, Status>> + Send + 'static

Server streaming response type for the DeltaAggregatedResources method.

Required Methods

fn stream_aggregated_resources<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<DiscoveryResponse>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamAggregatedResourcesStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

This RPC is not supported but remains here as part of the xDS standard, as Quilkin only uses a delta configuration transport for performance reasons.

fn delta_aggregated_resources<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<DeltaDiscoveryResponse>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DeltaAggregatedResourcesStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Delta (Incremental) xDS implementation